requirements.txt
streamlit
PyMuPDF
sumy
Pillow